Collapse ethereumservice into service
