+++ packages/client/src/service/fullethereumservice.ts
@@ -11,15 +11,14 @@ import { SnapProtocol } from '../net/protocol/snapprotocol'
 import { BeaconSynchronizer, FullSynchronizer, SnapSynchronizer } from '../sync'
 import { Skeleton } from '../sync/skeleton'
 
-import { EthereumService } from './ethereumservice'
+import { Service, type ServiceOptions } from './service'
 import { TxPool } from './txpool'
 
 import type { Peer } from '../net/peer/peer'
 import type { Protocol } from '../net/protocol'
-import type { EthereumServiceOptions } from './ethereumservice'
 import type { Block } from '@ethereumjs/block'
 
-interface FullEthereumServiceOptions extends EthereumServiceOptions {
+interface FullEthereumServiceOptions extends ServiceOptions {
   /** Serve LES requests (default: false) */
   lightserv?: boolean
 }
@@ -28,7 +27,7 @@ interface FullEthereumServiceOptions extends EthereumServiceOptions {
  * Full Ethereum service
  * @memberof module:service
  */
-export class FullEthereumService extends EthereumService {
+export class FullEthereumService extends Service {
   public synchronizer?: BeaconSynchronizer | FullSynchronizer | SnapSynchronizer
   public lightserv: boolean
   public miner: Miner | undefined
