+++ packages/client/src/service/index.ts
@@ -2,7 +2,6 @@
  * @module service
  */
 
-export * from './ethereumservice'
 export * from './fullethereumservice'
 export * from './lightethereumservice'
 export * from './service'
