+++ packages/client/src/service/service.ts
@@ -1,13 +1,36 @@
+import { Chain } from '../blockchain'
 import { PeerPool } from '../net/peerpool'
+import { FlowControl } from '../net/protocol'
 import { Event } from '../types'
+import { type V8Engine, getV8Engine } from '../util'
 
 import type { Config } from '../config'
 import type { Peer } from '../net/peer/peer'
 import type { Protocol } from '../net/protocol'
+import type { Synchronizer } from '../sync'
+import type { AbstractLevel } from 'abstract-level'
 
 export interface ServiceOptions {
   /* Config */
   config: Config
+
+  /* Blockchain */
+  chain: Chain
+
+  /* Blockchain database */
+  chainDB?: AbstractLevel<string | Uint8Array, string | Uint8Array, string | Uint8Array>
+
+  /* State database */
+  stateDB?: AbstractLevel<string | Uint8Array, string | Uint8Array, string | Uint8Array>
+
+  /* Meta database (receipts, logs, indexes) */
+  metaDB?: AbstractLevel<string | Uint8Array, string | Uint8Array, string | Uint8Array>
+
+  /* Sync retry interval in ms (default: 8000) */
+  interval?: number
+
+  /* Protocol timeout in ms (default: 6000) */
+  timeout?: number
 }
 
 /**
@@ -19,7 +42,27 @@ export class Service {
   public opened: boolean
   public running: boolean
   public pool: PeerPool
+  public flow: FlowControl
+  public chain: Chain
+  public interval: number
+  public timeout: number
+  public synchronizer?: Synchronizer
 
+  // A handle to v8Engine lib for mem stats, assigned on open if running in node
+  private v8Engine: V8Engine | null = null
+
+  /**
+   * Interval for client stats output (e.g. memory) (in ms)
+   */
+  private STATS_INTERVAL = 30000
+
+  /**
+   * Shutdown the client when memory threshold is reached (in percent)
+   *
+   */
+  private MEMORY_SHUTDOWN_THRESHOLD = 92
+
+  private _statsInterval: NodeJS.Timeout | undefined /* global NodeJS */
   /**
    * Create new service and associated peer pool
    */
@@ -45,6 +88,11 @@ export class Service {
       }
     })
 
+    this.flow = new FlowControl()
+    // @ts-ignore TODO replace with async create constructor
+    this.chain = options.chain ?? new Chain(options)
+    this.interval = options.interval ?? 8000
+    this.timeout = options.timeout ?? 6000
     this.opened = false
     this.running = false
   }
@@ -53,8 +101,7 @@ export class Service {
    * Service name
    */
   get name() {
-    return ''
-    //throw new Error('Unimplemented')
+    return 'eth'
   }
 
   /**
@@ -85,6 +132,8 @@ export class Service {
     )
 
     await this.pool.open()
+    await this.chain.open()
+    await this.synchronizer?.open()
     this.opened = true
     return true
   }
@@ -107,6 +156,16 @@ export class Service {
       return false
     }
     await this.pool.start()
+    void this.synchronizer?.start()
+    if (this.v8Engine === null) {
+      this.v8Engine = await getV8Engine()
+    }
+
+    this._statsInterval = setInterval(
+      // eslint-disable-next-line @typescript-eslint/await-thenable
+      await this.stats.bind(this),
+      this.STATS_INTERVAL
+    )
     this.running = true
     this.config.logger.info(`Started ${this.name} service.`)
     return true
@@ -118,11 +177,29 @@ export class Service {
   async stop(): Promise<boolean> {
     if (this.opened) {
       await this.close()
+      await this.synchronizer?.close()
     }
     await this.pool.stop()
+    clearInterval(this._statsInterval)
+    await this.synchronizer?.stop()
     this.running = false
     this.config.logger.info(`Stopped ${this.name} service.`)
     return true
+  }
+
+  stats() {
+    if (this.v8Engine !== null) {
+      const { used_heap_size, heap_size_limit } = this.v8Engine.getHeapStatistics()
+
+      const heapUsed = Math.round(used_heap_size / 1000 / 1000) // MB
+      const percentage = Math.round((100 * used_heap_size) / heap_size_limit)
+      this.config.logger.info(`Memory stats usage=${heapUsed} MB percentage=${percentage}%`)
+
+      if (percentage >= this.MEMORY_SHUTDOWN_THRESHOLD && !this.config.shutdown) {
+        this.config.logger.error('EMERGENCY SHUTDOWN DUE TO HIGH MEMORY LOAD...')
+        process.kill(process.pid, 'SIGINT')
+      }
+    }
   }
 
   /**
